pdf-converter/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
SQLAlchemy==2.0.21
python-docx==0.8.11
python-pptx==0.6.23
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pypdf==3.17.1
